app/chatapp/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.3.0
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.6
pydantic-settings==2.2.1
email-validator==2.1.1
httpx==0.27.0
